added burningship and color

+++ mandatory/src/colors.c
@@ -6,7 +6,7 @@
 /*   By: ycontre <ycontre@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/03 14:54:25 by ycontre           #+#    #+#             */
-/*   Updated: 2023/12/04 14:02:04 by ycontre          ###   ########.fr       */
+/*   Updated: 2023/12/04 15:00:54 by ycontre          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,7 +64,8 @@ int hsv_to_hex(double hue, double saturation, double value)
 int color_smoothing(double iteration, t_fractol *fractol)
 {
 	int max = 50;
-	long double test = (long double)iteration/(long double)max;
+	long double test = iteration/(long double)max;
+	
 	if (fractol->color_type == 0)
 		return (hsv_to_hex((int)(iteration * 15 + 150) % 360, 0.5, 1));
 	else if (fractol->color_type == 1)
